### 2023-06-30 #5607 @editorialbot
[PRE REVIEW]: Pinard: A Python package for Near Infrared Reflectance Spectroscopy
**Submitting author:** @gbeurier (<a href="http://orcid.org/0000-0002-5621-7079">Grégory Beurier</a>)
**Repository:** https://github.com/GBeurier/pinard
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.1
