### 2023-06-30 #5607 @Kevin-Mattheus-Moerman
@GBeurier thanks for this submission. I am the AEiC on this track and here to help with initial steps. For the moment, below are some points that require your attention:

- [ ] Please check that above potentially missing DOI, if the suggestion is correct please implement it. You can call `@editorialbot check references`, to check them again, and use `@editorialbot generate pdf` to update the paper draft.
- [ ] Please define the `NIRS` acronym in the text.
- [ ] Add space after full stop in `.Although`
