### 2023-07-10 772cfe3
Update paper.bib
Added missing doi.
