### 2023-07-10 #5607 @GBeurier
Hi @lucydot, thank you for reviewing pinard.
I have done the requested modifications and I suggest the following reviewers:
- [Ayoub Fatihi](https://reviewers.joss.theoj.org/reviewers/2933)
- [@ksunden](https://reviewers.joss.theoj.org/reviewers/559)
