+++ paper.bib
@@ -46,9 +46,7 @@
 	pages = {1-19},
 	year = {2007},
 	doi = {10.1255/jnirs.716},
-	URL = { 
-			https://doi.org/10.1255/jnirs.716
-	},
+	URL = {https://doi.org/10.1255/jnirs.716},
     abstract = { Science-based approaches to agricultural and environmental management are needed to accelerate development progress in the world's poorest countries. We present a diagnostic surveillance framework modelled on medical diagnostic approaches for evidence-based management of agriculture and environment in developing countries. Infrared (IR) spectroscopy can play a pivotal role in making the surveillance framework operational, by providing a rapid, low cost and highly reproducible diagnostic screening tool. We review the wide applicability of IR spectroscopy for setting up measurement systems for the management of soils, crops, agricultural inputs, livestock health, agricultural products and water quality. IR spectroscopy is already being used in the design of soil surveillance systems, but the principles are generally applicable. A new evidence-based interpretation approach to plant analysis, combining plant and soil IR spectroscopy measurements, is proposed. Finally, an idealised design is proposed for making IR spectroscopy-based diagnostic surveillance operational in developing countries over the next ten years. Large area surveillance frameworks for agricultural and environmental problems will deploy integrated spectral indicators of soil, crop and livestock health and water quality. Spectral indicators will help to quantify risk factors associated with problems and assess intervention impacts. Smallholder farmers will have access to IR spectroscopy-based analysis of soils, crops and inputs through a network of hand-held or mobile IR spectroscopy units. Agricultural processing industries will make extensive use of IR spectroscopy on the factory floor to add value to agricultural produce and improve food safety. Regional centres of scientific and technological excellence will be required to support (i) high quality laboratory reference analyses, (ii) development of IR spectroscopy calibration databases and interpretation systems and (iii) up-grading of scientific and technical skills through training and education. Key challenges for adoption of this design include (i) building human capacity in science- and technology-based approaches, (ii) development of rugged low cost IR spectroscopy instrumentation and (iii) development of decision support systems to interpret IR spectroscopy data into management recommendations. }
 }
 
@@ -71,7 +69,7 @@
 	number={4},
 	pages={767--778},
 	year={2020},
-	doi={https://doi.org/10.1111/pce.13718},
+	doi={10.1111/pce.13718},
 	publisher={Wiley Online Library}
 }
 
