### 2023-07-15 #5607 @ayoubft
Hey,
Thanks @lucydot for the welcome;
very excited to review for joss.
I will start by reading the docs.
