### 2023-08-10 #5747 @untzag
Hello @gbeurier, from the git history it seems like you are the only code contributor to Pinard. Could you please clarify how Denis Cornet and Lauriane Rouan contributed? The JOSS authorship guidelines might be relevant here.

https://joss.readthedocs.io/en/latest/submitting.html#authorship
