### 2023-08-10 #5747 @untzag
Hello @GBeurier, in my opinion your paper would be enhanced by an explicit "Impact" section. By explicitly discussing how Pinard has been used in past and ongoing projects, readers such as myself might get a better picture of how Pinard works, practically.

Currently you only say "Pinard has been successfully used in a number of scientific projects". As an outsider it's hard for me to see exactly what role Pinard plays in that paper.

This is just an opinion, I don't think an Impact section is required for acceptance to JOSS.
