### 2023-08-21 #5747 @lucydot
Hello, thanks for your early comments @untzag.

I agree that an Impact/Applications section is useful for code context. This is listed under "What your paper should my paper contain?" section of the JOSS submission docs - https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain

> Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.

The JOSS review process is iterative - so please feel free to respond to individual points @GBeurier as they arise rather than waiting for the full review to be complete. In my experience, an on-going conversation enables a better review process.  If you are unable to respond within a couple of weeks, then a timeframe is always useful.
