### 2023-11-16 cd485db
Naming and import fixes
Fix bug #14 - naming issue
Fix bug #16 and #20 - import was local on a wrongly named package. Tests updated.
Fix bug #17 - requirements updated
