+++ pinard/__init__.py
@@ -1,3 +1,3 @@
 __version__ = "1.0.1"
 
-__all__ = ["augmentation", "model_selection", "pipeline", "preprocessing"]
+__all__ = ["augmentation", "model_selection", "sklearn", "preprocessing"]
