+++ tests/test_global.py
@@ -1,17 +1,0 @@
-# Load data
-# split_data
-# Pipeline_XY:
-# [
-# 	augment
-# 	FeatureUnion#preprocess
-# 	scaler
-# ]
-
-# fit
-
-# Keras PipelineXY_validation:
-# [
-# 	SampleAugmentation
-# 	FeatureAugmentation#preprocess
-# ]
-# fit
