+++ pinard/sklearn/_pipeline.py
@@ -3,8 +3,6 @@ from joblib import Parallel, delayed
 from sklearn.pipeline import FeatureUnion, _fit_transform_one, _transform_one
 from sklearn.preprocessing import FunctionTransformer
 
-# from sklearn.utils.metaestimators import available_if
-
 
 class FeatureAugmentation(FeatureUnion):
     """Stacks results of multiple transformer objects in a new axis.
@@ -70,7 +68,7 @@ class SampleAugmentation(FeatureUnion):
         List of (name, trans) tuples specifying the transformer objects to be applied to the data.
     """
 
-    def __init__(self, transformer_list, *, n_jobs=None, transformer_weights=None, verbose=False):
+    def __init__(self, transformer_list, *, n_jobs=None, verbose=False):
         """
         This is the constructor for a class that initializes a list of transformers with optional
         parameters.
@@ -81,9 +79,6 @@ class SampleAugmentation(FeatureUnion):
         transformer. If it has three elements, the first element
         :param n_jobs: The number of CPU cores to use for parallel processing. If set to None, all
         available cores will be used
-        :param transformer_weights: A dictionary of weights assigned to each transformer. These weights
-        are used to compute the weighted average of the transformed features. If not provided, all
-        transformers are assumed to have equal weight
         :param verbose: A boolean parameter that controls whether or not progress messages are printed
         to the console during the fitting process. If set to True, progress messages will be printed. If
         set to False, no progress messages will be printed, defaults to False (optional)
@@ -103,7 +98,7 @@ class SampleAugmentation(FeatureUnion):
         super().__init__(
             transformer_origin_list,
             n_jobs=n_jobs,
-            transformer_weights=transformer_weights,
+            transformer_weights=None,
             verbose=verbose,
         )
 
@@ -153,120 +148,3 @@ class SampleAugmentation(FeatureUnion):
         Ys = np.repeat(y, self.total_count, axis=0)
 
         return Xs, Ys
-
-
-# def _transform_one(transformer, X, y, weight, **fit_params):
-#     res = transformer.transform(X)
-#     # if we have a weight for this transformer, multiply output
-#     if weight is None:
-#         return res
-#     return res * weight
-
-
-# def _transform_one_xy(transformer, X, y, weight, **fit_params):
-#     resX, resY = transformer.transform(X, y)
-#     # if we have a weight for this transformer, multiply output
-#     if weight is None:
-#         return resX, resY
-#     return resX * weight, resY
-
-
-# class Pipeline_XY(Pipeline):
-#     def _validate_steps(self):
-#         pass
-#         # names, estimators = zip(*self.steps)
-
-#         # # validate names
-#         # self._validate_names(names)
-
-#         # # validate estimators
-#         # transformers = estimators[:-1]
-#         # estimator = estimators[-1]
-
-#         # for t in transformers:
-#         #     if t is None or t == "passthrough":
-#         #         continue
-#         #     if not (hasattr(t, "fit") or hasattr(t, "fit_transform")) or not hasattr(
-#         #         t, "transform"
-#         #     ):
-#         #         raise TypeError(
-#         #             "All intermediate steps should be "
-#         #             "transformers and implement fit and transform "
-#         #             "or be the string 'passthrough' "
-#         #             "'%s' (type %s) doesn't" % (t, type(t))
-#         #         )
-
-#         # # We allow last estimator to be None as an identity transformation
-#         # if (
-#         #     estimator is not None
-#         #     and estimator != "passthrough"
-#         #     and not hasattr(estimator, "fit")
-#         # ):
-#         #     raise TypeError(
-#         #         "Last step of Pipeline should implement fit "
-#         #         "or be the string 'passthrough'. "
[…]
