+++ requirements.txt
@@ -4,6 +4,6 @@ numpy
 pandas
 pytest
 PyWavelets>=1.2.0
-setuptools>=65.0.0
 scikit-learn>=1.0.1
 scipy>=1.7.3
+twinning
