+++ docs/.readthedocs.yaml
@@ -1,0 +1,34 @@
+# .readthedocs.yaml
+version: 2
+
+# Build documentation in the docs/ directory with Sphinx
+sphinx:
+  configuration: docs/source/conf.py
+
+# Build documentation with MkDocs
+# mkdocs:
+#   configuration: mkdocs.yml
+
+# Optionally build your docs in additional formats such as PDF and ePub
+formats: all
+
+# Optionally set the version of Python and its dependencies
+python:
+  version: 3.7
+  install:
+    - requirements: docs/readthedocs.requirements.txt
+    - method: pip
+      path: .
+
+# Optionally declare the version of the tools you want to use
+build:
+  os: ubuntu-20.04
+  tools:
+    python: "3.7"
+    sphinx: "3.5.4"
+    mkdocs: "1.1.2"
+
+# Optionally set the submodules that you want to include in the build
+submodules:
+  include: all
+  recursive: true
