+++ docs/source/conf.py
@@ -23,6 +23,7 @@ extensions = [
     'sphinx.ext.napoleon',
     'sphinx.ext.autodoc',
     'sphinx_rtd_theme',
+    'sphinx_autodoc_typehints'
 ]
 
 templates_path = ['_templates']
