+++ docs/.readthedocs.yaml
@@ -19,5 +19,6 @@ build:
   tools:
     python: "3.7"
 
-sphinx:
-  configuration: docs/source/conf.py
+python:
+  install:
+    - requirements: "requirements.txt"
