+++ docs/source/conf.py
@@ -22,8 +22,7 @@ release = '1.0'
 extensions = [
     'sphinx.ext.napoleon',
     'sphinx.ext.autodoc',
-    'sphinx_rtd_theme',
-    'sphinx_autodoc_typehints'
+    'sphinx_rtd_theme'
 ]
 
 templates_path = ['_templates']
