+++ paper.bib
@@ -23,7 +23,7 @@
 	doi = {10.1038/s41597-023-02189-w},
 	abstract = {Data from functional trait databases have been increasingly used to address questions related to plant diversity and trait-environment relationships. However, such databases provide intraspecific data that combine individual records obtained from distinct populations at different sites and, hence, environmental conditions. This prevents distinguishing sources of variation (e.g., genetic-based variation vs. phenotypic plasticity), a necessary condition to test for adaptive processes and other determinants of plant phenotypic diversity. Consequently, individual traits measured under common growing conditions and encompassing within-species variation across the occupied geographic range have the potential to leverage trait databases with valuable data for functional and evolutionary ecology. Here, we recorded 16 functional traits and leaf hyperspectral reflectance (NIRS) data for 721 widely distributed Arabidopsis thaliana natural accessions grown in a common garden experiment. These data records, together with meteorological variables obtained during the experiment, were assembled to create the AraDiv dataset. AraDiv is a comprehensive dataset of A. thaliana’s intraspecific variability that can be explored to address questions at the interface of genetics and ecology.},
 	issn = {2052-4463},
-	url = {10.1038/s41597-023-02189-w},
+	url = {http://dx.doi.org/10.1038/s41597-023-02189-w},
 }
 
 @article{Vasseur2022Perspective,
