### 2023-11-23 8cf54b7
Augmentations tests
Pytest and validation files for augmentation
