+++ pinard/augmentation/_spline_augmentation.py
@@ -114,46 +114,228 @@ def interval_selection(n_l, CumVect):
     return np.min(i1), np.max(i2)
 
 
-class Random_X_Spline_Deformation(Augmenter):
+class Spline_Smoothing(Augmenter):
+    """
+    Class to apply a smoothing spline to a 1D signal.
+
+    Parameters
+    ----------
+    X : ndarray
+        Input data.
+    apply_on : str, optional
+        Apply augmentation on "samples" or "global" (default: "samples").
+    """
+
     def augment(self, X, apply_on="samples"):
         """
-        Randomly modify the x-coordinate based on subsampled spline.
+        Apply a smoothing spline to the data.
 
         Parameters
         ----------
         X : ndarray
             Input data.
         apply_on : str, optional
-            Apply augmentation on "samples" or "features" (default: "samples").
+            Apply augmentation on "samples" or "global" (default: "samples").
 
         Returns
         -------
         ndarray
             Augmented data.
         """
+        length = X.shape[-1]
+        x_abs = np.arange(length)
+        res = []
+        for x in X:
+            spl = interpolate.UnivariateSpline(x_abs, x, s=1 / length)
+            y_smooth = spl(x_abs)
+            res.append(y_smooth)
 
-        x = np.arange(0, len(X[0]), 1)
-        t, c, k = interpolate.splrep(x, X[0], s=0, k=3)
-        delta_x_size = int(np.around(len(t) / 20))
-        delta_x = np.linspace(np.min(x), np.max(x), delta_x_size)
-        delta_y = np.random.uniform(-10, 10, delta_x_size)
+        print(np.array(res).shape, X.shape)
+        return np.array(res)
+
+
+class Spline_X_Perturbations(Augmenter):
+    """
+    Class to apply a perturbation to a 1D signal using B-spline interpolation.
+
+    Parameters
+    ----------
+    X : ndarray
+        Input data.
+    apply_on : str, optional
+        Apply augmentation on "samples" or "global" (default: "samples").
+    spline_degree : int, optional
+        Degree of the spline. Default is 3 (cubic).
+    perturbation_density : float, optional
+        Density of perturbation points relative to data size. Default is 0.05.
+    perturbation_range : tuple, optional
+        Range of perturbation values (min, max). Default is (-10, 10).
+    """
+
+    def __init__(self, apply_on="samples", random_state=None, *, copy=True, spline_degree=3, perturbation_density=0.05, perturbation_range=(-10, 10)):
+        self.spline_degree = spline_degree
+        self.perturbation_density = perturbation_density
+        self.perturbation_range = perturbation_range
+        super().__init__(apply_on, random_state, copy=copy)
+
+    def augment(self, X, apply_on="samples"):
+        """
+        Augment the data with a perturbation using B-spline interpolation.
+
+        Parameters
+        ----------
+        X : ndarray
+            Input data to be augmented.
+        apply_on : str, optional
+            Apply augmentation on "samples" or "global" data. Default is "samples".
+
+        Returns
+        -------
+        ndarray
+            Augmented data.
+        """
+        if not 0 <= self.perturbation_density <= 1:
+            raise ValueError("Perturbation density must be between 0 and 1")
+
+        x_range = np.arange(X.shape[-1])
+        res = []
+
+        t, c, k = interpolate.splrep(x_range, X[0], s=0, k=self.spline_degree)
+        # Determine the number of perturbation points
+        delta_x_size = max(int(np.around(len(t) * self.perturbation_density)), 2)
+        delta_x = np.linspace(np.min(x_range), np.max(x_range), delta_x_size)
[…]
