+++ tests/conftest.py
@@ -1,10 +1,12 @@
 from os.path import abspath
 from os.path import dirname as d
+import operator
 
 import numpy as np
 import pytest
 
 from pinard import preprocessing as pp
+from pinard import augmentation as aug
 
 ROOT_DIR = d(d(abspath(__file__)))
 
@@ -76,3 +78,47 @@ def default_preprocessings():
         ("SimpleScale", pp.SimpleScale()),
     ]
     return preprocessings
+
+
+@pytest.fixture(scope="session")
+def augmentation_source_data():
+    csv_aug = np.loadtxt(
+        path_to("test_augmentation.csv"), delimiter=";")
+    return csv_aug[:, 1:]
+
+
+# @pytest.fixture(scope="session")
+# def augmenters():
+#     seed = 42
+#     return [
+#         (0, "Rotate_Translate", aug.Rotate_Translate, {"random_state": seed}),
+#         (1, "Rotate_Translate_custom", aug.Rotate_Translate, {"random_state": seed, "p_range": 5, "y_factor": 5}),
+#         (2, "Random_X_Operation", aug.Random_X_Operation, {"random_state": seed}),
+#         (3, "Random_X_Operation_custom", aug.Random_X_Operation, {"random_state": seed, "operator_func": operator.add, "operator_range": (-0.002, 0.002)}),
+#         (4, "Spline_Smoothing", aug.Spline_Smoothing, {"random_state": seed}),
+#         (5, "Spline_X_Perturbation", aug.Spline_X_Perturbations, {"random_state": seed}),
+#         (6, "Spline_X_Perturbation", aug.Spline_X_Perturbations, {"random_state": seed, "perturbation_density": 0.01, "perturbation_range": (-30, 30)}),
+#         (7, "Spline_Y_Perturbation", aug.Spline_Y_Perturbations, {"random_state": seed}),
+#         (8, "Spline_Y_Perturbation", aug.Spline_Y_Perturbations, {"random_state": seed, "spline_points": 5, "perturbation_intensity": 0.02}),
+#         (9, "Spline_X_Simplification", aug.Spline_X_Simplification, {"random_state": seed}),
+#         (10, "Spline_X_Simplification", aug.Spline_X_Simplification, {"random_state": seed, "spline_points": 15}),
+#         (11, "Spline_Curve_Simplification", aug.Spline_Curve_Simplification, {"random_state": seed}),
+#         (12, "Spline_Curve_Simplification", aug.Spline_Curve_Simplification, {"random_state": seed, "spline_points": 5, "uniform": False}),
+#         (13, "Spline_Curve_Simplification", aug.Spline_Curve_Simplification, {"random_state": seed, "spline_points": 5, "uniform": True}),
+#     ]
+
+# Fixture for loading augmented data from CSV files
+
+
+@pytest.fixture(scope="function")
+def augmented_global_data_csv(index, name):
+    data = np.loadtxt(
+        path_to(f"test_augmentation_{index}_{name}_global.csv"), delimiter=";")
+    return data
+
+
+@pytest.fixture(scope="function")
+def augmented_samples_data_csv(index, name):
+    data = np.loadtxt(
+        path_to(f"test_augmentation_{index}_{name}_samples.csv"), delimiter=";")
+    return data
