### 2023-11-23 8167156
Tmp utils refactoring and tests
Some comments cleaning too
