+++ pinard/augmentation/_spline_augmentation.py
@@ -435,41 +435,3 @@ class Spline_Curve_Simplification(Augmenter):
             simplified_X[i] = interpolate.BSpline(t, c, k, extrapolate=False)(x)
 
         return simplified_X
-
-
-# nfreq = X.shape[-1]
-# x0 = np.linspace(0, np.max(X[0]), nfreq)
-# res = X_length(x0, X[0])
-# nb_segments = 10
-
-# x_samples = []
-# y_samples = []
-
-# for s in range(1, nb_segments):
-#     length = X_length(x0, X[0])[0] / nb_segments
-#     n_l = s * length
-#     test = res[2]
-#     toto = interval_selection(n_l, test)
-
-#     P = segment_pt_coord(
-#         x1=x0[toto[1]],
-#         y1=X[toto[1]],
-#         x2=x0[toto[0]],
-#         y2=X[toto[0]],
-#         fracL=res[1][toto[1]] % length,
-#         L=res[1][toto[1]],
-#     )
-
-#     x_samples.append(P[0])
-#     y_samples.append(P[1])
-
-# x = np.array(x_samples)
-# x = np.concatenate(([0], x, [np.max(x0)]))
-# y = np.array(y_samples)
-# y = np.concatenate(([X[0][0]], y, [X[0][nfreq - 1]]))
-# t, c, k = interpolate.splrep(x, y, s=0, k=3)
-# xmin, xmax = x.min(), x.max()
-# xx = np.linspace(xmin, xmax, nfreq)
-# spline = interpolate.BSpline(t, c, k, extrapolate=False)
-
-# return spline(xx)
