+++ pinard/utils.py
@@ -1,3 +1,5 @@
+import logging
+import re
 import numpy as np
 import pandas as pd
 
@@ -118,3 +120,159 @@ def load_csv(
         raise WrongFormatError(x_data, y_data)
 
     return x_data, y_data
+
+
+# # Configure logging
+# logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+
+
+# def analyze_csv_file(filename, delimiter_candidates=None, numeric_delimiters=None):
+#     """
+#     Analyze a CSV file to determine its delimiter, numeric delimiter, and header presence.
+#     Parameters can be optionally set by the user.
+#     Raises exceptions for various error conditions.
+#     """
+#     delimiter_candidates = delimiter_candidates or [';', '\t', ',']
+#     numeric_delimiters = numeric_delimiters or ['.', ',']
+
+#     try:
+#         with open(filename, 'r', newline='') as file:
+#             first_lines = [next(file) for _ in range(5)]
+#     except Exception as e:
+#         logging.error(f"Error reading file: {filename} - {e}")
+#         raise
+
+#     delimiter_counts = {delimiter: sum(re.findall(re.escape(delimiter), line) for line in first_lines)
+#                         for delimiter in delimiter_candidates}
+#     if not any(delimiter_counts.values()):
+#         raise ValueError("No delimiter found in the file.")
+
+#     numeric_delimiter_counts = {numeric_delimiter: sum(re.findall(re.escape(numeric_delimiter), line) for line in first_lines[1:])
+#                                 for numeric_delimiter in numeric_delimiters}
+
+#     res_delimiter = max(delimiter_counts, key=delimiter_counts.get)
+#     res_numeric_delimiter = max(numeric_delimiter_counts, key=numeric_delimiter_counts.get)
+
+#     column_name_pattern = r'^[\w\d_]+$'
+#     column_header_present = all(re.match(column_name_pattern, cell.strip()) for cell in first_lines[0].split(res_delimiter))
+#     header = 0 if column_header_present else None
+
+#     return res_delimiter, res_numeric_delimiter, header
+
+
+# def load_csv(file_path, na_policy='auto', delimiter=None, header=None):
+#     """
+#     Load a CSV file with specified NA handling policy, delimiter, and header.
+#     """
+#     if not file_path.endswith('.csv'):
+#         raise ValueError("File format not supported. Please provide a CSV file.")
+
+#     try:
+#         delimiter, nb_delimiter, inferred_header = analyze_csv_file(file_path, delimiter_candidates=[delimiter] if delimiter else None)
+#         header = header if header is not None else inferred_header
+#         data = pd.read_csv(file_path, header=header, na_filter=False, sep=delimiter, engine='python', skip_blank_lines=False, decimal=nb_delimiter)
+#     except Exception as e:
+#         logging.error(f"Error loading file: {file_path} - {e}")
+#         raise
+
+#     data = data.replace(r"^\s*$", "", regex=True).applymap(lambda x: pd.to_numeric(x, errors='coerce'))
+
+#     data = data.dropna(how='all', axis=1).dropna(how='all', axis=0)
+#     rows_with_na = data.isna().any(axis=1)
+#     indexes_with_na = data[rows_with_na].index.tolist()
+
+#     if na_policy == 'abort':
+#         raise ValueError("NA values found, aborting as per policy.")
+#     elif na_policy == 'remove':
+#         data = data.drop(indexes_with_na)
+#     elif na_policy == 'replace':
+#         data = data.fillna(0)
+#     elif na_policy == 'auto':
[…]
