+++ tests/test_utils.py
@@ -1,6 +1,3 @@
-from data_loader import load_XY_csv, DatasetFormatError
-from data_loader import load_csv
-from data_loader import analyze_csv_file
 from pinard import utils
 import pytest
 
